pyproject.toml
[project]
name = "resume-automation"
version = "0.1.0"
description = "Resume automation system using AI and Google Drive integration"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi==0.104.1",
    "uvicorn==0.24.0",
    "pydantic==2.11.7",
    "python-multipart==0.0.6",
    "requests==2.31.0",
    "beautifulsoup4==4.12.2",
    "anthropic==0.40.0",
    "python-dotenv==1.0.0",
    "google-api-python-client==2.149.0",
    "google-auth==2.34.0",
    "google-auth-oauthlib==1.2.1",
    "google-auth-httplib2==0.2.0",
    "pyyaml==6.0.1",
    "python-docx==1.1.0",
    "docxtpl>=0.16.7",
    "pandas>=2.2.0",
    "openpyxl>=3.1.2",
]

[project.optional-dependencies]
dev = [
    "pytest==7.4.3",
    "pytest-asyncio==0.21.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["backend", "frontend"]

[tool.pytest.ini_options]
testpaths = ["backend/tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"